requirements.txt
# Django
Django==4.2.7

# Database
psycopg2-binary==2.9.7
dj-database-url==1.3.0

# Forms
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Utilities
django-humanize==0.1.2
Faker==19.6.2
python-decouple==3.8

# Static files
whitenoise==6.5.0

# Production
gunicorn==21.2.0

# Development
django-debug-toolbar==4.1.0